requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
Authlib==1.3.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
executing==2.1.0
faiss-cpu==1.9.0
fastapi==0.115.2
ffmpeg-python==0.2.0
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.10.0
future==1.0.0
gradio==5.3.0
gradio_client==1.4.2
greenlet==3.1.1
grpcio==1.67.0
grpcio-health-checking==1.67.0
grpcio-tools==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.0
huggingface-hub==0.26.1
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.3.4
langchain-core==0.3.12
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langchainhub==0.1.21
langsmith==0.1.137
llama-cloud==0.1.4
llama-index==0.11.19
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.19
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-index-vector-stores-weaviate==1.1.2
llama-parse==0.5.11
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
openai==1.52.0
openai-whisper==20240930
opencv-python==4.10.0.84
orjson==3.10.9
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.1.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
ruff==0.7.0
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.40.0
striprtf==0.0.26
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tomlkit==0.12.0
torch==2.2.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
validators==0.34.0
wcwidth==0.2.13
weaviate-client==4.9.0
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.16.0
